pyproject.toml
[project]
name = "idefics-nnx"
version = "0.0.1"
description = "Flax NNX Implementation of Idefics family of VLMs by 🤗"
readme = ".github/README.md"
requires-python = ">=3.10"
license = { file = "LICENSE" }
dependencies = [
  "einops",
  "flax>=0.9.0",
  "jaxtyping"
]

[project.optional-dependencies]
test = [
  "pytest>=6.0.0",
  "torch",
  "transformers",
]
dev = [
  "ruff"
]

[tool.ruff]
line-length = 90

[tool.ruff.lint]
select = ["E", "F", "I001"]

[tool.ruff.lint.isort]
lines-after-imports = 2

[tool.pytest.ini_options]
testpaths = ["tests"]
filterwarnings = [
  "ignore::DeprecationWarning",
	"ignore::UserWarning"
]
